+++ src/pages/Search/SearchPageNarrow/index.tsx
@@ -1,13 +1,10 @@
 import React, {useCallback, useState} from 'react';
 import {useSharedValue} from 'react-native-reanimated';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
-import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import NavigationTabBar from '@components/Navigation/NavigationTabBar';
 import NAVIGATION_TABS from '@components/Navigation/NavigationTabBar/NAVIGATION_TABS';
 import ScreenWrapper from '@components/ScreenWrapper';
 import {useSearchContext} from '@components/Search/SearchContext';
-import type {SearchHeaderOptionValue} from '@components/Search/SearchPageHeader/SearchPageHeader';
-import type {BankAccountMenuItem, SearchQueryJSON} from '@components/Search/types';
 import useHandleBackButton from '@hooks/useHandleBackButton';
 import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
@@ -22,26 +19,7 @@ import {turnOffMobileSelectionMode} from '@userActions/MobileSelectionMode';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
-import type {SearchResults} from '@src/types/onyx';
-import type {PaymentMethodType} from '@src/types/onyx/OriginalMessage';
-import type {SearchResultsInfo} from '@src/types/onyx/SearchResults';
-
-type SearchPageNarrowProps = {
-    queryJSON?: SearchQueryJSON;
-    metadata?: SearchResultsInfo;
-    headerButtonsOptions: Array<DropdownOption<SearchHeaderOptionValue>>;
-    searchResults?: SearchResults;
-    isMobileSelectionModeEnabled: boolean;
-    footerData: {
-        count: number | undefined;
-        total: number | undefined;
-        currency: string | undefined;
-    };
-    currentSelectedPolicyID?: string | undefined;
-    currentSelectedReportID?: string | undefined;
-    confirmPayment?: (paymentType: PaymentMethodType | undefined) => void;
-    latestBankItems?: BankAccountMenuItem[] | undefined;
-};
+import type SearchPageNarrowProps from './types';
 
 function SearchPageNarrow({
     queryJSON,
